Transition to WSL

+++ core/imagelib/text.py
@@ -22,7 +22,7 @@ def get_text_image( shape, text, color=(1,1,1), border=0.2, font=None):
         canvas = Image.new('RGB', (w,h) , (0,0,0) )
         draw = ImageDraw.Draw(canvas)
         offset = ( 0, 0)
-        draw.text(offset, text, font=pil_font, fill=tuple((np.array(color)*255).astype(np.int)) )
+        draw.text(offset, text, font=pil_font, fill=tuple((np.array(color)*255).astype(np.int32)) )
 
         result = np.asarray(canvas) / 255
 
@@ -59,6 +59,6 @@ def draw_text_lines (image, rect, text_lines, color=(1,1,1), border=0.2, font=No
         draw_text (image, (l, i*h_per_line, r, (i+1)*h_per_line), text_lines[i], color, border, font)
 
 def get_draw_text_lines ( image, rect, text_lines, color=(1,1,1), border=0.2, font=None):
-    image = np.zeros ( image.shape, dtype=np.float )
+    image = np.zeros ( image.shape, dtype=np.float32 )
     draw_text_lines ( image, rect, text_lines, color, border, font)
     return image
