+++ facelib/LandmarksProcessor.py
@@ -324,14 +324,15 @@ def get_transform_mat (image_landmarks, output_size, face_type, scale=1.0):
     if not remove_align:
         l_t = np.array( [ g_c - tb_diag_vec*mod,
                           g_c + bt_diag_vec*mod,
-                          g_c + tb_diag_vec*mod ] )
+                          g_c + tb_diag_vec*mod
+                          ], dtype=np.float32 )
     else:
         # remove_align - face will be centered in the frame but not aligned
         l_t = np.array( [ g_c - tb_diag_vec*mod,
                           g_c + bt_diag_vec*mod,
                           g_c + tb_diag_vec*mod,
                           g_c - bt_diag_vec*mod,
-                         ] )
+                         ], dtype=np.float32 )
 
         # get area of face square in global space
         area = mathlib.polygon_area(l_t[:,0], l_t[:,1] )
@@ -342,7 +343,8 @@ def get_transform_mat (image_landmarks, output_size, face_type, scale=1.0):
         # calc 3 points with unrotated square
         l_t = np.array( [ g_c + [-side,-side],
                           g_c + [ side,-side],
-                          g_c + [ side, side] ] )
+                          g_c + [ side, side]
+                          ], dtype=np.float32 )
 
     # calc affine transform from 3 global space points to 3 local space points size of 'output_size'
     pts2 = np.float32(( (0,0),(output_size,0),(output_size,output_size) ))
@@ -361,7 +363,7 @@ def get_rect_from_landmarks(image_landmarks):
 def expand_eyebrows(lmrks, eyebrows_expand_mod=1.0):
     if len(lmrks) != 68:
         raise Exception('works only with 68 landmarks')
-    lmrks = np.array( lmrks.copy(), dtype=np.int )
+    lmrks = np.array( lmrks.copy(), dtype=np.int32 )
 
     # #nose
     ml_pnt = (lmrks[36] + lmrks[0]) // 2
@@ -416,7 +418,7 @@ def get_image_eye_mask (image_shape, image_landmarks):
 
     hull_mask = np.zeros( (h,w,1),dtype=np.float32)
 
-    image_landmarks = image_landmarks.astype(np.int)
+    image_landmarks = image_landmarks.astype(np.int32)
 
     cv2.fillConvexPoly( hull_mask, cv2.convexHull( image_landmarks[36:42]), (1,) )
     cv2.fillConvexPoly( hull_mask, cv2.convexHull( image_landmarks[42:48]), (1,) )
@@ -439,7 +441,7 @@ def get_image_mouth_mask (image_shape, image_landmarks):
 
     hull_mask = np.zeros( (h,w,1),dtype=np.float32)
 
-    image_landmarks = image_landmarks.astype(np.int)
+    image_landmarks = image_landmarks.astype(np.int32)
 
     cv2.fillConvexPoly( hull_mask, cv2.convexHull( image_landmarks[60:]), (1,) )
 
@@ -688,7 +690,7 @@ def draw_landmarks (image, image_landmarks, color=(0,255,0), draw_circles=True, 
     if len(image_landmarks) != 68:
         raise Exception('get_image_eye_mask works only with 68 landmarks')
 
-    int_lmrks = np.array(image_landmarks, dtype=np.int)
+    int_lmrks = np.array(image_landmarks, dtype=np.int32)
 
     jaw = int_lmrks[slice(*landmarks_68_pt["jaw"])]
     right_eyebrow = int_lmrks[slice(*landmarks_68_pt["right_eyebrow"])]
